+++ plugins/org.jkiss.dbeaver.ext.db2.ui/OSGI-INF/l10n/bundle_ru.properties
@@ -1,0 +1,20 @@
+# DBeaver - Universal Database Manager
+# Copyright (C) 2013-2016 Denis Forveille titou10.titou10@gmail.com
+# Copyright (C) 2010-2020 DBeaver Corp and others serge@jkiss.org
+
+Bundle-Vendor = JKISS
+Bundle-Name = DBeaver DB2 UI
+
+dialog.connection.header=DB2 Connection Settings
+
+editor.source.ddl.name=DDL
+editor.source.ddl.description=DDL
+
+tool.org.jkiss.dbeaver.ext.db2.table.reorg.name=Reorg Table...
+tool.org.jkiss.dbeaver.ext.db2.table.reorgix.name=Reorg Index...
+tool.org.jkiss.dbeaver.ext.db2.table.reorgcheck.name=Reorg Check (Table)...
+tool.org.jkiss.dbeaver.ext.db2.table.reorgcheckix.name=Reorg Check (Indexes)...
+tool.org.jkiss.dbeaver.ext.db2.table.setintegrity.name=Set Integrity...
+tool.org.jkiss.dbeaver.ext.db2.table.runstats.name=Runstats...
+tool.org.jkiss.dbeaver.ext.db2.table.truncate.name=Truncate...
+tool.org.jkiss.dbeaver.ext.db2.showerror.name=Retrieve SQL Error Message...
