+++ plugins/org.jkiss.dbeaver.ext.db2.ui/build.properties
@@ -1,0 +1,8 @@
+source.. = src/
+output.. = target/classes/
+bin.includes = plugin.xml,\
+               META-INF/,\
+               OSGI-INF/,\
+               .,\
+               icons/,\
+               templates/
