+++ plugins/org.jkiss.dbeaver.ext.db2.ui/src/org/jkiss/dbeaver/ext/db2/ui/tools/maintenance/DB2RunstatsTool.java
@@ -1,0 +1,45 @@
+/*
+ * DBeaver - Universal Database Manager
+ * Copyright (C) 2013-2015 Denis Forveille (titou10.titou10@gmail.com)
+ * Copyright (C) 2010-2020 DBeaver Corp and others
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.jkiss.dbeaver.ext.db2.ui.tools.maintenance;
+
+import org.eclipse.ui.IWorkbenchPart;
+import org.eclipse.ui.IWorkbenchWindow;
+import org.jkiss.dbeaver.DBException;
+import org.jkiss.dbeaver.ext.db2.model.DB2Table;
+import org.jkiss.dbeaver.model.struct.DBSObject;
+import org.jkiss.dbeaver.ui.tools.IUserInterfaceTool;
+import org.jkiss.utils.CommonUtils;
+
+import java.util.Collection;
+import java.util.List;
+
+/**
+ * DB2 table runstats action
+ */
+public class DB2RunstatsTool implements IUserInterfaceTool {
+
+    @Override
+    public void execute(IWorkbenchWindow window, IWorkbenchPart activePart, Collection<DBSObject> objects) throws DBException
+    {
+        List<DB2Table> tables = CommonUtils.filterCollection(objects, DB2Table.class);
+        if (!tables.isEmpty()) {
+            DB2RunstatsDialog dialog = new DB2RunstatsDialog(activePart.getSite(), tables);
+            dialog.open();
+        }
+    }
+}
