+++ plugins/org.jkiss.dbeaver.ext.db2/META-INF/MANIFEST.MF
@@ -7,20 +7,16 @@ Bundle-Version: 2.1.117.qualifier
 Bundle-Release-Date: 20200504
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-ActivationPolicy: lazy
-Bundle-Activator: org.jkiss.dbeaver.ext.db2.Activator
 Require-Bundle: org.eclipse.core.runtime,
  org.eclipse.jface,
  org.eclipse.jface.text,
- org.eclipse.ui,
  org.eclipse.ui.views,
  org.eclipse.ui.editors,
  org.eclipse.ui.workbench.texteditor,
  org.jkiss.dbeaver.model,
  org.jkiss.dbeaver.registry,
  org.jkiss.dbeaver.ui,
- org.jkiss.dbeaver.tasks.ui,
  org.jkiss.dbeaver.ui.editors.base,
- org.jkiss.dbeaver.ui.editors.connection,
  org.jkiss.dbeaver.ui.editors.sql,
  org.jkiss.dbeaver.ui.editors.session
 Bundle-ClassPath: .
